requirements_frozen.txt
aiohttp==3.9.1
aiosignal==1.4.0
aiosqlite==0.19.0
alembic==1.13.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
Authlib==1.6.5
azure-ai-formrecognizer==3.3.2
azure-common==1.1.28
azure-core==1.29.5
azure-identity==1.15.0
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
billiard==4.2.2
black==23.12.1
blis==0.7.11
build==1.3.0
cachetools==6.2.1
catalogue==2.0.10
celery==5.3.6
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
coverage==7.11.0
cryptography==46.0.3
cymem==2.0.11
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
ecdsa==0.19.1
email-validator==2.1.0.post1
et_xmlfile==2.0.0
fastapi==0.109.0
filelock==3.20.0
flake8==7.0.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
google-auth==2.42.1
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-health-checking==1.76.0
grpcio-tools==1.76.0
gunicorn==21.2.0
h11==0.16.0
hf-xet==1.2.0
hiredis==2.3.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.26.0
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
isodate==0.7.2
isort==5.13.2
Jinja2==3.1.3
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.4
kubernetes==34.1.0
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.23
langchain-openai==0.0.5
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.0.87
language_data==1.3.0
limits==5.6.0
lxml==5.1.0
Mako==1.3.10
marisa-trie==1.3.1
Markdown==3.5.2
MarkupSafe==3.0.3
marshmallow==3.26.1
mccabe==0.7.0
mmh3==5.2.0
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.7.0
murmurhash==1.0.13
mypy==1.8.0
mypy_extensions==1.1.0
neo4j==5.16.0
networkx==3.5
nltk==3.8.1
numpy==1.26.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
onnxruntime==1.23.2
openai==1.10.0
openpyxl==3.1.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
overrides==7.7.0
packaging==23.2
pandas==2.1.4
passlib==1.7.4
pathspec==0.12.1
pdf2image==1.17.0
pillow==10.2.0
pinecone-client==3.0.2
platformdirs==4.5.0
pluggy==1.6.0
posthog==6.7.11
preshed==3.0.10
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psycopg2-binary==2.9.9
pulsar-client==3.8.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyflakes==3.2.0
PyJWT==2.8.0
pypdf==4.0.1
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytesseract==0.3.10
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-pptx==0.6.23
pytz==2025.2
PyYAML==6.0.1
redis==5.0.1
regex==2025.10.23
reportlab==4.4.4
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==2.3.1
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
slowapi==0.1.9
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.8
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.25
srsly==2.5.1
starlette==0.35.1
sympy==1.14.0
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.6.0
tiktoken==0.5.2
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
triton==3.5.0
typer==0.9.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.27.0
uvloop==0.22.1
validators==0.22.0
vine==5.1.0
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.2.14
weasel==0.3.4
weaviate-client==4.4.0
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
xlsxwriter==3.2.9
yarl==1.22.0
zipp==3.23.0
